fixes views/peewee
